1. database.migration package revised. 2. MetaDataAdapter deleted because it is no longer required. 3. custom_print module modified not to print to output in migration mode if force=False. 4. custom_print enhanced.

+++ src/pyrin/database/migration/hooks.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 """
-database hooks module.
+database migration hooks module.
 """
 
 import pyrin.database.migration.services as migration_services
 import pyrin.configuration.services as config_services
+import pyrin.application.services as application_services
 
 from pyrin.application.decorators import application_hook
 from pyrin.application.hooks import ApplicationHookBase
@@ -53,20 +54,17 @@ class ApplicationHook(ApplicationHookBase):
         this method will be called after application has been loaded.
         """
 
-        # we should check whether it is required to drop
-        # all entities in database on server startup.
-        if config_services.get('database', 'migration', 'drop_on_startup') is True:
-            environment = config_services.get_active('environment', 'env')
-            debug = config_services.get_active('environment', 'debug')
-            unit_testing = config_services.get_active('environment', 'unit_testing')
+        if application_services.is_migration() is False:
+            if config_services.get('database', 'migration', 'drop_on_startup') is True:
+                environment = config_services.get_active('environment', 'env')
+                debug = config_services.get_active('environment', 'debug')
+                unit_testing = config_services.get_active('environment', 'unit_testing')
 
-            if (environment == 'development' and debug is True) or \
-                    (environment == 'testing' and unit_testing is True):
-                print_warning('Dropping all models...')
-                migration_services.drop_all()
+                if (environment == 'development' and debug is True) or \
+                        (environment == 'testing' and unit_testing is True):
+                    print_warning('Dropping all models...')
+                    migration_services.drop_all()
 
-        # we should check whether it is required to create
-        # all entities in database on server startup.
-        if config_services.get('database', 'migration', 'create_on_startup') is True:
-            print_info('Creating all models...')
-            migration_services.create_all()
+            if config_services.get('database', 'migration', 'create_on_startup') is True:
+                print_info('Creating all models...')
+                migration_services.create_all()
